+++ DatadogSessionReplay/Tests/Mocks/RecorderMocks.swift
@@ -396,11 +396,9 @@ extension RUMContext: AnyMockable, RandomMockable {
 
     public static func mockRandom() -> RUMContext {
         return RUMContext(
-            ids: .init(
-                applicationID: .mockRandom(),
-                sessionID: .mockRandom(),
-                viewID: .mockRandom()
-            ),
+            applicationID: .mockRandom(),
+            sessionID: .mockRandom(),
+            viewID: .mockRandom(),
             viewServerTimeOffset: .mockRandom()
         )
     }
@@ -412,11 +410,9 @@ extension RUMContext: AnyMockable, RandomMockable {
         serverTimeOffset: TimeInterval = .mockAny()
     ) -> RUMContext {
         return RUMContext(
-            ids: .init(
-                applicationID: applicationID,
-                sessionID: sessionID,
-                viewID: viewID
-            ),
+            applicationID: applicationID,
+            sessionID: sessionID,
+            viewID: viewID,
             viewServerTimeOffset: serverTimeOffset
         )
     }
@@ -454,9 +450,9 @@ extension Recorder.Context: AnyMockable, RandomMockable {
     ) {
         self.init(
             privacy: privacy,
-            applicationID: rumContext.ids.applicationID,
-            sessionID: rumContext.ids.sessionID,
-            viewID: rumContext.ids.viewID ?? "",
+            applicationID: rumContext.applicationID,
+            sessionID: rumContext.sessionID,
+            viewID: rumContext.viewID ?? "",
             viewServerTimeOffset: rumContext.viewServerTimeOffset,
             date: date
         )