+++ DatadogTrace/Tests/ContextMessageReceiverTests.swift
@@ -6,6 +6,7 @@
 
 import XCTest
 import TestUtilities
+import DatadogInternal
 
 @testable import DatadogTrace
 
@@ -30,58 +31,135 @@ class ContextMessageReceiverTests: XCTestCase {
     func testItReceivesRUMContext() throws {
         // Given
         let receiver = ContextMessageReceiver(bundleWithRumEnabled: true)
-        var ids: [String: String?] = [
-            "application.id": "app-id",
-            "session.id": "session-id",
-            "view.id": "view-id",
-            "user_action.id": "action-id"
-        ]
+        let core = PassthroughCoreMock()
 
-        let core = PassthroughCoreMock(
-            context: .mockWith(featuresAttributes: ["rum": ["ids": ids]]),
-            messageReceiver: receiver
+        let coreContext1: DatadogContext = .mockWith(
+            baggages: [
+                "rum": .init([
+                    "application.id": "app-id",
+                    "session.id": "session-id",
+                    "view.id": "view-id",
+                    "user_action.id": "action-id"
+                ])
+            ]
         )
 
-        var rumContext = try XCTUnwrap(receiver.context.rum)
-        XCTAssertEqual(rumContext["_dd.application.id"] as? String, "app-id")
-        XCTAssertEqual(rumContext["_dd.session.id"] as? String, "session-id")
-        XCTAssertEqual(rumContext["_dd.view.id"] as? String, "view-id")
-        XCTAssertEqual(rumContext["_dd.action.id"] as? String, "action-id")
+        let coreContext2: DatadogContext = .mockWith(
+            baggages: [
+                "rum": .init([
+                    "application.id": "app-id",
+                    "session.id": "session-id",
+                    "view.id": nil,
+                    "user_action.id": nil
+                ])
+            ]
+        )
 
         // When
-        ids = [
-            "application.id": "app-id",
-            "session.id": "session-id",
-            "view.id": nil,
-            "user_action.id": nil
-        ]
-        core.set(feature: "rum", attributes: { ["ids": ids] })
+        XCTAssert(
+            receiver.receive(message: .context(coreContext1), from: core)
+        )
 
         // Then
-        rumContext = try XCTUnwrap(receiver.context.rum)
-        XCTAssertEqual(rumContext["_dd.application.id"] as? String, "app-id")
-        XCTAssertEqual(rumContext["_dd.session.id"] as? String, "session-id")
-        XCTAssertNil(rumContext["_dd.view.id"] as Any?)
-        XCTAssertNil(rumContext["_dd.action.id"] as Any?)
+        XCTAssertEqual(receiver.context.rum?["_dd.application.id"], "app-id")
+        XCTAssertEqual(receiver.context.rum?["_dd.session.id"], "session-id")
+        XCTAssertEqual(receiver.context.rum?["_dd.view.id"], "view-id")
+        XCTAssertEqual(receiver.context.rum?["_dd.action.id"], "action-id")
+
+        // When
+        XCTAssert(
+            receiver.receive(message: .context(coreContext2), from: core)
+        )
+
+        // Then
+        XCTAssertEqual(receiver.context.rum?["_dd.application.id"], "app-id")
+        XCTAssertEqual(receiver.context.rum?["_dd.session.id"], "session-id")
+        XCTAssertNil(receiver.context.rum?["_dd.view.id"])
+        XCTAssertNil(receiver.context.rum?["_dd.action.id"])
+    }
+
+    func testItReceivesNilRUMContext() throws {
+        // Given
+        let receiver = ContextMessageReceiver(bundleWithRumEnabled: true)
+        let core = PassthroughCoreMock()
+
+        let coreContext: DatadogContext = .mockWith(
+            baggages: [
+                "rum": .init([
+                    "application.id": "app-id",
+                    "session.id": "session-id",
+                    "view.id": "view-id",
+                    "user_action.id": "action-id"
+                ])
+            ]
+        )
+
+        // When
+        XCTAssert(
+            receiver.receive(message: .context(coreContext), from: core)
+        )
+
+        // Then
+        XCTAssertEqual(receiver.context.rum?["_dd.application.id"], "app-id")
+        XCTAssertEqual(receiver.context.rum?["_dd.session.id"], "session-id")
+        XCTAssertEqual(receiver.context.rum?["_dd.view.id"], "view-id")
+        XCTAssertEqual(receiver.context.rum?["_dd.action.id"], "action-id")
+
+        // When
+        XCTAssert(
+            receiver.receive(message: .context(.mockAny()), from: core)
+        )
+
+        // Then
+        XCTAssertNil(receiver.context.rum)
+    }
+
+    func testItReceivesMalformedRUMContext() throws {
+        // Given
+        let telemetryReceiver = TelemetryReceiverMock()
+        let receiver = ContextMessageReceiver(bundleWithRumEnabled: true)
+        let core = PassthroughCoreMock(
+            messageReceiver: telemetryReceiver
+        )
+
+        let coreContext: DatadogContext = .mockWith(
+            baggages: [
+                "rum": .init("malformed RUM context")
+            ]
+        )
+
+        // When
+        XCTAssert(
+            receiver.receive(message: .context(coreContext), from: core)
+        )
+
+        // Then
+        XCTAssertNil(receiver.context.rum)
+
+        let error = try XCTUnwrap(telemetryReceiver.messages.first?.asError)
+        XCTAssert(error.message.contains("Fails to decode RUM context from Trace - typeMismatch"))
     }
 
     func testItIngnoresRUMContext() throws {
         // Given
         let receiver = ContextMessageReceiver(bundleWithRumEnabled: false)
-        var ids: [String: String?] = .mockRandom()
-
-        let core = PassthroughCoreMock(
-            context: .mockWith(featuresAttributes: ["rum": ["ids": ids]]),
-            messageReceiver: receiver
+        let core = PassthroughCoreMock()
+        let coreContext: DatadogContext = .mockWith(
+            baggages: [
+                "rum": .init([
+                    "application.id": "app-id",
+                    "session.id": "session-id",
+                    "view.id": "view-id",
+                    "user_action.id": "action-id"
+                ])
+            ]
         )
 
-        XCTAssertNil(receiver.context.rum)
-
         // When
-        ids = .mockRandom()
-        core.set(feature: "rum", attributes: { ["ids": ids] })
+        XCTAssert(
+            receiver.receive(message: .context(coreContext), from: core)
+        )
 
-        // Then
         XCTAssertNil(receiver.context.rum)
     }
 }