+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Unreleased
 
-# 2.3.0 / 27-09-2023
+# 2.4.0 / 18-10-2023
+
+- [FEATURE] WebView Log events can be now sampled. See [#1515][]
+- [BUGFIX] WebView RUM events are now dropped if mobile RUM session is not sampled. See [#1502][]
+- [BUGFIX] Fix `os.name` in Log events. See [#1493][]
+
+# 2.3.0 / 02-10-2023
 
 - [IMPROVEMENT] Add UIBackgroundTask for uploading jobs. See [#1412][]
 - [IMPROVEMENT] Report Build Number in Logs and RUM. See [#1465][]
@@ -533,8 +539,11 @@ Release `2.0` introduces breaking changes. Follow the [Migration Guide](MIGRATIO
 [#1464]: https://github.com/DataDog/dd-sdk-ios/pull/1464
 [#1412]: https://github.com/DataDog/dd-sdk-ios/pull/1412
 [#1488]: https://github.com/DataDog/dd-sdk-ios/pull/1488
+[#1502]: https://github.com/DataDog/dd-sdk-ios/pull/1502
+[#1515]: https://github.com/DataDog/dd-sdk-ios/pull/1515
 [#1465]: https://github.com/DataDog/dd-sdk-ios/pull/1465
 [#1498]: https://github.com/DataDog/dd-sdk-ios/pull/1498
+[#1493]: https://github.com/DataDog/dd-sdk-ios/pull/1493
 [@00fa9a]: https://github.com/00FA9A
 [@britton-earnin]: https://github.com/Britton-Earnin
 [@hengyu]: https://github.com/Hengyu